+++ packages/apollo-angular-components/projects/angular/components/ui-grid/src/managers/resize/types/resizableGrid.ts
@@ -27,4 +27,7 @@ export abstract class ResizableGrid<T> {
 
     protected abstract _destroyed$: Subject<void>;
     protected abstract _columnChanges$: Observable<SimpleChanges>;
+    abstract columnWidthPercentToPxRatio: number;
 }
+
+export type ResizeEmission = Record<string, { initialPercentage: number; finalPercentage: number }>;